library and notebooks refactoring
