--- gdeep/neural_nets/__init__.py
@@ -1,15 +1,0 @@
-
-from .sample_nn import SimpleNN, DeeperNN, LogisticRegressionNN, Net
-from .utility import train_classification_nn, SaveOutput, SaveNodeOutput, SaveLayerOutput, Layers_list, get_activations
-
-__all__ = [
-    'SimpleNN',
-    'DeeperNN',
-    'LogisticRegressionNN',
-    'Net',
-    'train_classification_nn',
-    'SaveOutput',
-    'SaveNodeOutput',
-    'SaveLayerOutput',
-    'get_activations'
-    ]
