+++ gdeep/plotting/__init__.py
@@ -1,0 +1,13 @@
+
+from .pca_activations import plot_PCA_activations
+from .plot_decision_bdry import plot_decision_boundary, plot_activation_contours
+from .persistence_activations import plot_persistence_diagrams, \
+                                     persistence_diagrams_of_activations
+
+__all__ = [
+    'plot_PCA_activations',
+    'plot_persistence_diagrams',
+    'plot_decision_boundary',
+    'plot_activation_contours',
+    'persistence_diagrams_of_activations'
+    ]
