+++ local_test.bh
@@ -2,7 +2,7 @@ python -m pip install --upgrade pip
 
 pip install -e .
 
-pip install pytest nbconvert jupyter flake8
+pip install pytest nbconvert jupyter flake8 --upgrade
 
 # stop the build if there are Python syntax errors or undefined names
 flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
@@ -14,5 +14,5 @@ pytest
 set -e
 for n in examples/*.ipynb
 do
-  jupyter nbconvert --ExecutePreprocessor.timeout 500 --to notebook --execute $n
+  jupyter nbconvert --to notebook --execute $n
 done
